Refactored MultipleEntitySelect and splitted into 2 components
