+++ front/src/modules/ui/components/menu/DropdownMenu.tsx
@@ -17,6 +17,4 @@ export const DropdownMenu = styled.div`
   height: fit-content;
 
   width: 200px;
-
-  z-index: ${(props) => props.theme.lastLayerZIndex};
 `;
